+++ client/src/main/java/org/apache/celeborn/client/ShuffleClientImpl.java
@@ -24,13 +24,10 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 import scala.reflect.ClassTag$;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.util.concurrent.Uninterruptibles;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.CompositeByteBuf;
 import io.netty.buffer.Unpooled;
 import org.slf4j.Logger;
@@ -48,14 +45,10 @@ import org.apache.celeborn.common.network.client.RpcResponseCallback;
 import org.apache.celeborn.common.network.client.TransportClient;
 import org.apache.celeborn.common.network.client.TransportClientFactory;
 import org.apache.celeborn.common.network.protocol.PushData;
-import org.apache.celeborn.common.network.protocol.PushDataHandShake;
 import org.apache.celeborn.common.network.protocol.PushMergedData;
-import org.apache.celeborn.common.network.protocol.RegionFinish;
-import org.apache.celeborn.common.network.protocol.RegionStart;
 import org.apache.celeborn.common.network.server.BaseMessageHandler;
 import org.apache.celeborn.common.network.util.TransportConf;
 import org.apache.celeborn.common.protocol.*;
-import org.apache.celeborn.common.protocol.message.ControlMessages;
 import org.apache.celeborn.common.protocol.message.ControlMessages.*;
 import org.apache.celeborn.common.protocol.message.StatusCode;
 import org.apache.celeborn.common.rpc.RpcAddress;
@@ -72,7 +65,7 @@ import org.apache.celeborn.common.write.PushState;
 public class ShuffleClientImpl extends ShuffleClient {
   private static final Logger logger = LoggerFactory.getLogger(ShuffleClientImpl.class);
 
-  private static final byte MASTER_MODE = PartitionLocation.Mode.MASTER.mode();
+  protected static final byte MASTER_MODE = PartitionLocation.Mode.MASTER.mode();
 
   private static final Random RND = new Random();
 
@@ -85,24 +78,24 @@ public class ShuffleClientImpl extends ShuffleClient {
   private int maxReviveTimes;
   private boolean testRetryRevive;
   private final int pushBufferMaxSize;
-  private final long pushDataTimeout;
+  protected final long pushDataTimeout;
 
   private final RpcEnv rpcEnv;
 
-  private RpcEndpointRef driverRssMetaService;
+  protected RpcEndpointRef driverRssMetaService;
 
   protected TransportClientFactory dataClientFactory;
 
-  final int BATCH_HEADER_SIZE = 4 * 4;
+  protected final int BATCH_HEADER_SIZE = 4 * 4;
 
   // key: shuffleId, value: (partitionId, PartitionLocation)
   private final Map<Integer, ConcurrentHashMap<Integer, PartitionLocation>> reducePartitionMap =
       new ConcurrentHashMap<>();
 
-  private final ConcurrentHashMap<Integer, Set<String>> mapperEndMap = new ConcurrentHashMap<>();
+  protected final ConcurrentHashMap<Integer, Set<String>> mapperEndMap = new ConcurrentHashMap<>();
 
   // key: shuffleId-mapId-attemptId
-  private final Map<String, PushState> pushStates = new ConcurrentHashMap<>();
+  protected final Map<String, PushState> pushStates = new ConcurrentHashMap<>();
 
   private final ExecutorService pushDataRetryPool;
 
@@ -140,8 +133,6 @@ public class ShuffleClientImpl extends ShuffleClient {
 
   // key: shuffleId
   protected final Map<Integer, ReduceFileGroups> reduceFileGroupsMap = new ConcurrentHashMap<>();
-
-  private ConcurrentHashMap<String, TransportClient> currentClient = new ConcurrentHashMap<>();
 
   public ShuffleClientImpl(CelebornConf conf, UserIdentifier userIdentifier) {
     super();
@@ -455,7 +446,7 @@ public class ShuffleClientImpl extends ShuffleClient {
     return null;
   }
 
-  private void limitMaxInFlight(String mapKey, PushState pushState, String hostAndPushPort)
+  protected void limitMaxInFlight(String mapKey, PushState pushState, String hostAndPushPort)
       throws IOException {
     boolean reachLimit = pushState.limitMaxInFlight(hostAndPushPort);
 
@@ -465,7 +456,7 @@ public class ShuffleClientImpl extends ShuffleClient {
     }
   }
 
-  private void limitZeroInFlight(String mapKey, PushState pushState) throws IOException {
+  protected void limitZeroInFlight(String mapKey, PushState pushState) throws IOException {
     boolean reachLimit = pushState.limitZeroInFlight();
 
     if (reachLimit) {
@@ -1291,9 +1282,6 @@ public class ShuffleClientImpl extends ShuffleClient {
       pushState.exception.compareAndSet(null, new CelebornIOException("Cleaned Up"));
       pushState.cleanup();
     }
-    if (currentClient != null) {
-      currentClient.remove(mapKey);
-    }
   }
 
   @Override
@@ -1457,7 +1445,7 @@ public class ShuffleClientImpl extends ShuffleClient {
     driverRssMetaService = endpointRef;
   }
 
-  private boolean mapperEnded(int shuffleId, int mapId, int attemptId) {
+  protected boolean mapperEnded(int shuffleId, int mapId, int attemptId) {
     return mapperEndMap.containsKey(shuffleId)
         && mapperEndMap.get(shuffleId).contains(Utils.makeMapKey(shuffleId, mapId, attemptId));
   }
@@ -1496,388 +1484,5 @@ public class ShuffleClientImpl extends ShuffleClient {
     return (message.startsWith("Connection from ") && message.endsWith(" closed"))
         || (message.equals("Connection reset by peer"))
         || (message.startsWith("Failed to send RPC "));
-  }
-
-  public int pushDataToLocation(
-      String applicationId,
-      int shuffleId,
-      int mapId,
-      int attemptId,
-      int partitionId,
-      ByteBuf data,
-      PartitionLocation location,
-      Runnable closeCallBack)
-      throws IOException {
-    // mapKey
-    final String mapKey = Utils.makeMapKey(shuffleId, mapId, attemptId);
-    final String shuffleKey = Utils.makeShuffleKey(applicationId, shuffleId);
-    // return if shuffle stage already ended
-    if (mapperEnded(shuffleId, mapId, attemptId)) {
-      logger.info(
-          "Push data byteBuf to location {} ignored because mapper already ended for shuffle {} map {} attempt {}.",
-          location.hostAndPushPort(),
-          shuffleId,
-          mapId,
-          attemptId);
-      PushState pushState = pushStates.get(mapKey);
-      if (pushState != null) {
-        pushState.cleanup();
-      }
-      return 0;
-    }
-
-    PushState pushState = getPushState(mapKey);
-
-    // increment batchId
-    final int nextBatchId = pushState.nextBatchId();
-    int totalLength = data.readableBytes();
-    data.markWriterIndex();
-    data.writerIndex(0);
-    data.writeInt(partitionId);
-    data.writeInt(attemptId);
-    data.writeInt(nextBatchId);
-    data.writeInt(totalLength - BATCH_HEADER_SIZE);
-    data.resetWriterIndex();
-    logger.debug(
-        "Do push data byteBuf size {} for app {} shuffle {} map {} attempt {} reduce {} batch {}.",
-        totalLength,
-        applicationId,
-        shuffleId,
-        mapId,
-        attemptId,
-        partitionId,
-        nextBatchId);
-    // check limit
-    limitMaxInFlight(mapKey, pushState, location.hostAndPushPort());
-
-    // add inFlight requests
-    pushState.addBatch(nextBatchId, location.hostAndPushPort());
-
-    // build PushData request
-    NettyManagedBuffer buffer = new NettyManagedBuffer(data);
-    PushData pushData = new PushData(MASTER_MODE, shuffleKey, location.getUniqueId(), buffer);
-
-    // build callback
-    RpcResponseCallback callback =
-        new RpcResponseCallback() {
-          @Override
-          public void onSuccess(ByteBuffer response) {
-            pushState.removeBatch(nextBatchId, location.hostAndPushPort());
-            if (response.remaining() > 0) {
-              byte reason = response.get();
-              if (reason == StatusCode.STAGE_ENDED.getValue()) {
-                mapperEndMap
-                    .computeIfAbsent(shuffleId, (id) -> ConcurrentHashMap.newKeySet())
-                    .add(mapKey);
-              }
-            }
-            logger.debug(
-                "Push data byteBuf to {} success for shuffle {} map {} attemptId {} batch {}.",
-                location.hostAndPushPort(),
-                shuffleId,
-                mapId,
-                attemptId,
-                nextBatchId);
-          }
-
-          @Override
-          public void onFailure(Throwable e) {
-            pushState.removeBatch(nextBatchId, location.hostAndPushPort());
-            if (pushState.exception.get() != null) {
-              return;
-            }
-            if (!mapperEnded(shuffleId, mapId, attemptId)) {
-              String errorMsg =
-                  String.format(
-                      "Push data byteBuf to %s failed for shuffle %d map %d attempt %d batch %d.",
-                      location.hostAndPushPort(), shuffleId, mapId, attemptId, nextBatchId);
-              pushState.exception.compareAndSet(null, new CelebornIOException(errorMsg, e));
-            } else {
-              logger.warn(
-                  "Push data to {} failed but mapper already ended for shuffle {} map {} attempt {} batch {}.",
-                  location.hostAndPushPort(),
-                  shuffleId,
-                  mapId,
-                  attemptId,
-                  nextBatchId);
-            }
-          }
-        };
-    // do push data
-    try {
-      TransportClient client = createClientWaitingInFlightRequest(location, mapKey, pushState);
-      client.pushData(pushData, pushDataTimeout, callback, closeCallBack);
-    } catch (Exception e) {
-      logger.error(
-          "Exception raised while pushing data byteBuf for shuffle {} map {} attempt {} partitionId {} batch {} location {}.",
-          shuffleId,
-          mapId,
-          attemptId,
-          partitionId,
-          nextBatchId,
-          location,
-          e);
-      callback.onFailure(
-          new CelebornIOException(StatusCode.PUSH_DATA_CREATE_CONNECTION_FAIL_MASTER, e));
-    }
-    return totalLength;
-  }
-
-  private TransportClient createClientWaitingInFlightRequest(
-      PartitionLocation location, String mapKey, PushState pushState)
-      throws IOException, InterruptedException {
-    TransportClient client =
-        dataClientFactory.createClient(
-            location.getHost(), location.getPushPort(), location.getId());
-    if (currentClient.get(mapKey) != client) {
-      // makesure that messages have been sent by old client, in order to keep receiving data
-      // orderly
-      if (currentClient.get(mapKey) != null) {
-        limitZeroInFlight(mapKey, pushState);
-      }
-      currentClient.put(mapKey, client);
-    }
-    return currentClient.get(mapKey);
-  }
-
-  @Override
-  public void pushDataHandShake(
-      String applicationId,
-      int shuffleId,
-      int mapId,
-      int attemptId,
-      int numPartitions,
-      int bufferSize,
-      PartitionLocation location)
-      throws IOException {
-    final String mapKey = Utils.makeMapKey(shuffleId, mapId, attemptId);
-    final PushState pushState = pushStates.computeIfAbsent(mapKey, (s) -> new PushState(conf));
-    sendMessageInternal(
-        shuffleId,
-        mapId,
-        attemptId,
-        location,
-        pushState,
-        () -> {
-          String shuffleKey = Utils.makeShuffleKey(applicationId, shuffleId);
-          logger.info(
-              "PushDataHandShake shuffleKey {} attemptId {} locationId {}",
-              shuffleKey,
-              attemptId,
-              location.getUniqueId());
-          logger.debug("PushDataHandShake location {}", location.toString());
-          TransportClient client = createClientWaitingInFlightRequest(location, mapKey, pushState);
-          PushDataHandShake handShake =
-              new PushDataHandShake(
-                  MASTER_MODE,
-                  shuffleKey,
-                  location.getUniqueId(),
-                  attemptId,
-                  numPartitions,
-                  bufferSize);
-          client.sendRpcSync(handShake.toByteBuffer(), conf.pushDataTimeoutMs());
-          return null;
-        });
-  }
-
-  @Override
-  public Optional<PartitionLocation> regionStart(
-      String applicationId,
-      int shuffleId,
-      int mapId,
-      int attemptId,
-      PartitionLocation location,
-      int currentRegionIdx,
-      boolean isBroadcast)
-      throws IOException {
-    final String mapKey = Utils.makeMapKey(shuffleId, mapId, attemptId);
-    final PushState pushState = pushStates.computeIfAbsent(mapKey, (s) -> new PushState(conf));
-    return sendMessageInternal(
-        shuffleId,
-        mapId,
-        attemptId,
-        location,
-        pushState,
-        () -> {
-          String shuffleKey = Utils.makeShuffleKey(applicationId, shuffleId);
-          logger.info(
-              "RegionStart for shuffle {} regionId {} attemptId {} locationId {}.",
-              shuffleId,
-              currentRegionIdx,
-              attemptId,
-              location.getUniqueId());
-          logger.debug("RegionStart  for location {}.", location.toString());
-          TransportClient client = createClientWaitingInFlightRequest(location, mapKey, pushState);
-          RegionStart regionStart =
-              new RegionStart(
-                  MASTER_MODE,
-                  shuffleKey,
-                  location.getUniqueId(),
-                  attemptId,
-                  currentRegionIdx,
-                  isBroadcast);
-          ByteBuffer regionStartResponse =
-              client.sendRpcSync(regionStart.toByteBuffer(), conf.pushDataTimeoutMs());
-          if (regionStartResponse.hasRemaining()
-              && regionStartResponse.get() == StatusCode.HARD_SPLIT.getValue()) {
-            // if split then revive
-            PbChangeLocationResponse response =
-                driverRssMetaService.askSync(
-                    ControlMessages.Revive$.MODULE$.apply(
-                        applicationId,
-                        shuffleId,
-                        mapId,
-                        attemptId,
-                        location.getId(),
-                        location.getEpoch(),
-                        location,
-                        StatusCode.HARD_SPLIT),
-                    conf.requestPartitionLocationRpcAskTimeout(),
-                    ClassTag$.MODULE$.apply(PbChangeLocationResponse.class));
-            // per partitionKey only serve single PartitionLocation in Client Cache.
-            StatusCode respStatus = Utils.toStatusCode(response.getStatus());
-            if (StatusCode.SUCCESS.equals(respStatus)) {
-              return Optional.of(PbSerDeUtils.fromPbPartitionLocation(response.getLocation()));
-            } else if (StatusCode.MAP_ENDED.equals(respStatus)) {
-              mapperEndMap
-                  .computeIfAbsent(shuffleId, (id) -> ConcurrentHashMap.newKeySet())
-                  .add(mapKey);
-              return Optional.empty();
-            } else {
-              // throw exception
-              logger.error(
-                  "Exception raised while reviving for shuffle {} map {} attemptId {} partition {} epoch {}.",
-                  shuffleId,
-                  mapId,
-                  attemptId,
-                  location.getId(),
-                  location.getEpoch());
-              throw new CelebornIOException("RegionStart revive failed");
-            }
-          }
-          return Optional.empty();
-        });
-  }
-
-  @Override
-  public void regionFinish(
-      String applicationId, int shuffleId, int mapId, int attemptId, PartitionLocation location)
-      throws IOException {
-    final String mapKey = Utils.makeMapKey(shuffleId, mapId, attemptId);
-    final PushState pushState = pushStates.computeIfAbsent(mapKey, (s) -> new PushState(conf));
-    sendMessageInternal(
-        shuffleId,
-        mapId,
-        attemptId,
-        location,
-        pushState,
-        () -> {
-          final String shuffleKey = Utils.makeShuffleKey(applicationId, shuffleId);
-          logger.info(
-              "RegionFinish for shuffle {} map {} attemptId {} locationId {}.",
-              shuffleId,
-              mapId,
-              attemptId,
-              location.getUniqueId());
-          logger.debug("RegionFinish for location {}.", location.toString());
-          TransportClient client = createClientWaitingInFlightRequest(location, mapKey, pushState);
-          RegionFinish regionFinish =
-              new RegionFinish(MASTER_MODE, shuffleKey, location.getUniqueId(), attemptId);
-          client.sendRpcSync(regionFinish.toByteBuffer(), conf.pushDataTimeoutMs());
-          return null;
-        });
-  }
-
-  private <R> R sendMessageInternal(
-      int shuffleId,
-      int mapId,
-      int attemptId,
-      PartitionLocation location,
-      PushState pushState,
-      ThrowingExceptionSupplier<R, Exception> supplier)
-      throws IOException {
-    int batchId = 0;
-    try {
-      // mapKey
-      final String mapKey = Utils.makeMapKey(shuffleId, mapId, attemptId);
-      // return if shuffle stage already ended
-      if (mapperEnded(shuffleId, mapId, attemptId)) {
-        logger.debug(
-            "Send message to {} ignored because mapper already ended for shuffle {} map {} attempt {}.",
-            location.hostAndPushPort(),
-            shuffleId,
-            mapId,
-            attemptId);
-        return null;
-      }
-      pushState = getPushState(mapKey);
-      // force data has been send
-      limitZeroInFlight(mapKey, pushState);
-
-      // add inFlight requests
-      batchId = pushState.nextBatchId();
-      pushState.addBatch(batchId, location.hostAndPushPort());
-      return retrySendMessage(supplier);
-    } finally {
-      if (pushState != null) {
-        pushState.removeBatch(batchId, location.hostAndPushPort());
-      }
-    }
-  }
-
-  @FunctionalInterface
-  interface ThrowingExceptionSupplier<R, E extends Exception> {
-    R get() throws E;
-  }
-
-  private <R> R retrySendMessage(ThrowingExceptionSupplier<R, Exception> supplier)
-      throws IOException {
-
-    int retryTimes = 0;
-    boolean isSuccess = false;
-    Exception currentException = null;
-    R result = null;
-    while (!Thread.currentThread().isInterrupted()
-        && !isSuccess
-        && retryTimes < conf.networkIoMaxRetries(TransportModuleConstants.PUSH_MODULE)) {
-      logger.debug("RetrySendMessage  retry times {}.", retryTimes);
-      try {
-        result = supplier.get();
-        isSuccess = true;
-      } catch (Exception e) {
-        currentException = e;
-        if (e instanceof InterruptedException) {
-          Thread.currentThread().interrupt();
-        }
-        if (shouldRetry(e)) {
-          retryTimes++;
-          Uninterruptibles.sleepUninterruptibly(
-              conf.networkIoRetryWaitMs(TransportModuleConstants.PUSH_MODULE),
-              TimeUnit.MILLISECONDS);
-        } else {
-          break;
-        }
-      }
-    }
-    if (!isSuccess) {
-      if (currentException instanceof IOException) {
-        throw (IOException) currentException;
-      } else {
-        throw new CelebornIOException(currentException.getMessage(), currentException);
-      }
-    }
-    return result;
-  }
-
-  private boolean shouldRetry(Throwable e) {
-    boolean isIOException =
-        e instanceof IOException
-            || e instanceof TimeoutException
-            || (e.getCause() != null && e.getCause() instanceof TimeoutException)
-            || (e.getCause() != null && e.getCause() instanceof IOException)
-            || (e instanceof RuntimeException
-                && e.getMessage() != null
-                && e.getMessage().startsWith(IOException.class.getName()));
-    return isIOException;
   }
 }
